Move LessRunes into main
This comparitor function could hardly be considored a type. It's also
very small and probably not too useful overall so keep it in utils.
